fix: bound vMerge planning across the whole pass, and plan one merge per head row
Round-six review: two criticals, both introduced by this PR.

The probe allowance is one pool for the whole layout pass again, inherited by
every nested table. A probe re-enters nested-table layout, which plans the level
below and probes again, so a per-table allowance let each of sixteen levels
re-spend the whole thing: measured at three rows and two merges per level, depth
6 ran 63ms, depth 10 ran 4.3s, and `MAX_TABLE_NESTING` extrapolated past an hour
from a document of a few kilobytes. A file controls that nest, so the ceiling has
to bound the document rather than each table. The pool makes a table's plan
depend on what the pass walked before it, which is the trade the resolve budget
beside it already makes; an unbounded document is not on the table.

Only one merge per head row is planned. Detaching a second one empties the head
row of the last thing in it, so the row collapses and the first span's rows stop
adding up to the height it was sized at — its content painted below the table.
Span heights are also derived when the row is placed rather than stored when the
merge is admitted, so nothing that changes afterwards leaves a height behind.

The one-pass callers measure a merged head at the row's real top, like the
paginator, and place the row again unplanned if a head still comes back
incomplete — they have no next page to carry a remainder to. The `w:cantSplit`
whole-row move re-offers spans like every other break that retries its row. The
plan reads `tableLeft` through a getter, because a positioned probe localizes
wrap bands against it and both header placement and continuation re-derive it.
A shape the plan will decline no longer charges the probe budget first.

`table-fragment-finalize.ts` takes fragment finalize out of row layout, which
had reached its line cap again.
