+++ packages/core/src/layout/semantic-table-layout.ts
@@ -21,9 +21,7 @@
 import type { OoxmlElement, OoxmlNode } from '@docx-editor.dev/core/store';
 import {
   clipInlineDrawingRecordToRegion,
-  shiftInlineDrawingRecord,
   publishAnchoredDrawingsForParagraph,
-  anchoredDrawingAtomsInParagraph,
   type AnchoredDrawingRecord,
   type DrawingAnchorFrameContext,
 } from './drawing-layout.ts';
@@ -85,27 +83,18 @@ import type {
 import { firstLineShift, type ResolvedListItem } from './list-resolve.ts';
 import { publishListMarker } from './list-marker.ts';
 import { annotateTableFragmentGeometry } from './semantic-table-interaction.ts';
-import {
-  borderExtentPt,
-  resolveTableCellBorderGrid,
-  type BorderGridCell,
-  type BorderGridGeometry,
-  type TableBorderBox,
-  type TableBorderOwnershipBudget,
-} from './table-borders.ts';
-import {
-  resolveVMergeSpans,
-  type TableVMergeResolveBudget,
-  type TableVMergeResolveWork,
-} from './table-vmerge.ts';
-import {
-  acceptVMergeSpansAt,
-  createVMergePlanBudget,
-  planTableVMergeHeights,
-} from './table-vmerge-heights.ts';
+import { borderExtentPt, type TableBorderOwnershipBudget } from './table-borders.ts';
+import { type TableVMergeResolveBudget } from './table-vmerge.ts';
+import { acceptVMergeSpansAt, planTableVMergeHeights } from './table-vmerge-heights.ts';
 import { contentInsets } from './table-cell-geometry.ts';
+import { finalizeTableRows, shiftBlocks } from './table-fragment-finalize.ts';
+export { finalizeTableRows } from './table-fragment-finalize.ts';
 export { rowWithSplitBorders } from './table-cell-geometry.ts';
-import type { RowVMergeLayoutOptions, VMergeRowHeights } from './table-vmerge-heights.ts';
+import type {
+  RowVMergeLayoutOptions,
+  VMergePlanBudget,
+  VMergeRowHeights,
+} from './table-vmerge-heights.ts';
 
 export {
   createTableBorderOwnershipBudget,
@@ -153,9 +142,7 @@ export function paragraphDocumentOrderOf(
 export {
   createTableVMergeResolveBudget,
   MAX_VMERGE_RESOLVE_CELLS,
-  resolveVMergeSpans,
   type TableVMergeResolveBudget,
-  type TableVMergeResolveWork,
 } from './table-vmerge.ts';
 
 /** Soft ceiling on fragments emitted for one authored row (hostile / runaway splits). */
@@ -246,6 +233,15 @@ export interface TableFlowDeps {
    * pass. Exhaustion fails soft (remaining restarts keep rowSpan 1).
    */
   readonly vMergeResolveBudget?: TableVMergeResolveBudget;
+  /**
+   * What the vMerge height plan may spend across this PASS: cell visits on the authored
+   * rows, and probe layouts. Nested tables inherit the same object, which is the point —
+   * probes re-enter nested layout, so an allowance that does not compose across nesting
+   * bounds no document. Not the budget above: finalize spends that one, and sharing it
+   * exhausts it at half the cell count, where finalize fails soft to `rowSpan` 1 and
+   * leaves holes in the painted grid.
+   */
+  readonly vMergePlanBudget?: VMergePlanBudget;
   /**
    * Which tracked revisions this pass resolves away. A cell paragraph must resolve the same
    * mode as a body paragraph, or one table would show the proposed result while the text
@@ -388,48 +384,6 @@ function publishDeferredRowAnchors(
   }
 }
 
-function republishAnchoredParagraphsInBlocks(
-  blocks: readonly BlockFragmentRecord[],
-  authoredBlocks: readonly OoxmlElement[],
-  cellBox: LayoutBox,
-  deps: TableFlowDeps
-): void {
-  if (
-    !deps.onAnchorRepublish ||
-    !deps.inlineDrawingLayout ||
-    !deps.anchorFrameBase ||
-    !deps.pageContentClip
-  ) {
-    return;
-  }
-  for (const block of blocks) {
-    if (block.kind !== 'paragraph') continue;
-    const paragraph = authoredBlocks.find(
-      (candidate) => candidate.kind === 'paragraph' && candidate.id === block.paragraphId
-    );
-    if (!paragraph || paragraph.kind !== 'paragraph') continue;
-    const atoms = anchoredDrawingAtomsInParagraph(paragraph, deps.inlineDrawingLayout);
-    if (atoms.length === 0) continue;
-    deps.onAnchorRepublish(
-      block.paragraphId,
-      publishAnchoredDrawingsForParagraph({
-        paragraph,
-        paragraphId: block.paragraphId,
-        paragraphBox: block.box,
-        lines: block.lines,
-        drawingLayout: deps.inlineDrawingLayout,
-        frameBase: deps.anchorFrameBase(),
-        columnBox: deps.columnBoxForParagraph?.(block.box) ?? block.box,
-        cellBox,
-        pageClip: deps.pageContentClip(),
-        measurer: deps.measurer,
-        ...(deps.layoutTextboxStoryFor ? { layoutTextboxStory: deps.layoutTextboxStoryFor } : {}),
-        ...(deps.displayMode ? { displayMode: deps.displayMode } : {}),
-      })
-    );
-  }
-}
-
 function rowDepsForAnchors(
   deps: TableFlowDeps,
   deferredRowAnchors: DeferredRowAnchor[]
@@ -474,73 +428,6 @@ function sumCols(cols: readonly number[], from: number, to: number): number {
   let sum = 0;
   for (let index = from; index < to && index < cols.length; index += 1) sum += cols[index]!;
   return sum;
-}
-
-function shiftBlocks(blocks: readonly BlockFragmentRecord[], dy: number): BlockFragmentRecord[] {
-  if (dy === 0) return [...blocks];
-  return blocks.map((block) => {
-    if (block.kind === 'table') {
-      return {
-        ...block,
-        box: { ...block.box, y: block.box.y + dy },
-        rows: block.rows.map((row) => ({
-          ...row,
-          box: { ...row.box, y: row.box.y + dy },
-          cells: row.cells.map((cell) => ({
-            ...cell,
-            box: { ...cell.box, y: cell.box.y + dy },
-            blocks: shiftBlocks(cell.blocks, dy),
-          })),
-        })),
-      };
-    }
-    return {
-      ...block,
-      box: { ...block.box, y: block.box.y + dy },
-      ...(block.shadingBox
-        ? { shadingBox: { ...block.shadingBox, y: block.shadingBox.y + dy } }
-        : {}),
-      ...(block.bottomBorder
-        ? {
-            bottomBorder: {
-              ...block.bottomBorder,
-              box: { ...block.bottomBorder.box, y: block.bottomBorder.box.y + dy },
-            },
-          }
-        : {}),
-      // Every `w:pBdr` stroke, not only the bottom one. vAlign moves the whole paragraph
-      // down its cell; a frame left at the pre-shift y would sit above the text it encloses.
-      ...(block.borders
-        ? {
-            borders: block.borders.map((strokeRecord) => ({
-              ...strokeRecord,
-              box: { ...strokeRecord.box, y: strokeRecord.box.y + dy },
-            })),
-          }
-        : {}),
-      ...(block.marker
-        ? {
-            marker: {
-              ...block.marker,
-              box: { ...block.marker.box, y: block.marker.box.y + dy },
-            },
-          }
-        : {}),
-      lines: block.lines.map((line) => ({
-        ...line,
-        box: { ...line.box, y: line.box.y + dy },
-        spans: line.spans.map((span) => ({
-          ...span,
-          box: { ...span.box, y: span.box.y + dy },
-        })),
-        ...(line.drawings
-          ? {
-              drawings: line.drawings.map((drawing) => shiftInlineDrawingRecord(drawing, 0, dy)),
-            }
-          : {}),
-      })),
-    };
-  });
 }
 
 /**
@@ -1576,13 +1463,56 @@ export function measureRowHeight(
 }
 
 /**
+ * One row of a table placed in ONE PASS, where there is no next page to carry a remainder.
+ *
+ * `acceptVMergeSpansAt` measures a merged head at this row's real top with wrap bands
+ * applied, so the head's bound is what placement produces and the row comes back complete.
+ * If it ever does not, the row is placed again with the merge unplanned: the head then sizes
+ * its own row and holds all of its content, which is what this module did before it planned
+ * anything. Dropping the remainder instead would lose the lines silently.
+ */
+function layoutOnePassRow(
+  row: SemanticTableRow,
+  structure: SemanticTableStructure,
+  left: number,
+  y: number,
+  depth: number,
+  deps: TableFlowDeps,
+  isHeaderRepeat: boolean,
+  vMerge: RowVMergeLayoutOptions | undefined
+): LayoutRowBoundedResult {
+  const placed = layoutRowFragment(
+    row,
+    structure.columnWidthsPt,
+    left,
+    y,
+    isHeaderRepeat,
+    depth,
+    deps,
+    structure.cellSpacingPt,
+    vMerge
+  );
+  if (placed.remainder === null || vMerge === undefined) return placed;
+  return layoutRowFragment(
+    row,
+    structure.columnWidthsPt,
+    left,
+    y,
+    isHeaderRepeat,
+    depth,
+    deps,
+    structure.cellSpacingPt
+  );
+}
+
+/**
  * The vMerge plan for one table, probed by row layout. `rows` narrows it to the rows the
  * caller actually places — the paginator plans over the BODY rows, so a merge is never
  * planned against a repeated header copy.
  */
 export const vMergePlanFor = (
   structure: SemanticTableStructure,
-  left: number,
+  left: number | (() => number),
   depth: number,
   deps: TableFlowDeps,
   rows?: readonly SemanticTableRow[]
@@ -1593,159 +1523,11 @@ export const vMergePlanFor = (
     depth,
     deps,
     measureRowHeight,
-    // A budget PER TABLE, not per pass. A shared pool makes a table's height model depend on
-    // how many tables a pass happened to walk before it, so a resumed pass that starts lower
-    // in the document can size a merge as a span where a full pass falls back to row sizing —
-    // a reflow nothing the user did explains. Each table still gets a hard ceiling.
-    createVMergePlanBudget()
+    // The PASS's pool, inherited by every nested table through `deps`. A fresh allowance per
+    // table lets each nesting level re-spend the whole thing, which is exponential in depth
+    // on a nest a file controls.
+    deps.vMergePlanBudget
   );
-
-/**
- * After all rows of a table fragment are placed: expand vMerge restart boxes, re-apply
- * vAlign over the full span, and publish collapsed border edges.
- */
-export function finalizeTableRows(
-  rows: readonly TableRowFragmentRecord[],
-  structure: SemanticTableStructure,
-  sourceRows: readonly SemanticTableRow[],
-  ownershipBudget?: TableBorderOwnershipBudget,
-  vMergeBudget?: TableVMergeResolveBudget,
-  vMergeWork?: TableVMergeResolveWork,
-  onAnchorShift?: (paragraphId: string, dy: number) => void,
-  anchorDeps?: TableFlowDeps
-): TableRowFragmentRecord[] {
-  if (rows.length === 0) return [];
-
-  // Map laid-out cells back to authored structure cells (same order within each row).
-  const authoredById = new Map<string, SemanticTableCell>();
-  for (const row of sourceRows) {
-    for (const cell of row.cells) authoredById.set(cell.id, cell);
-  }
-
-  // One-pass column-keyed merge spans — O(cells), not O(rows × columns²). These rows are
-  // one PAGE FRAGMENT: a merge whose restart was placed on an earlier page is headed here
-  // by its continuation copy, which is why that copy comes back keyed in the span map.
-  const mergeSpanById = resolveVMergeSpans(rows, vMergeWork, vMergeBudget, {
-    pageFragment: true,
-  });
-
-  // Expand restart heights and shift content for vAlign over the full span.
-  const expanded: TableRowFragmentRecord[] = rows.map((row, rowIndex) => ({
-    ...row,
-    cells: row.cells.map((cell) => {
-      const resolvedSpan = mergeSpanById.get(cell.id);
-      if (cell.vMergeContinue && resolvedSpan === undefined) {
-        return { ...cell, paintInert: true, rowSpan: 1, borders: {}, blocks: [] };
-      }
-      // A carried-in continuation paints like the restart it continues: Word draws the
-      // merged cell's rules, fill and box on every page the merge crosses. Its content
-      // stayed with the restart on the earlier page, so `blocks` is empty either way.
-      const carried = cell.vMergeContinue;
-      const span = resolvedSpan ?? 1;
-      let height = cell.box.height;
-      if (span > 1) {
-        const last = rows[rowIndex + span - 1]!;
-        height = last.box.y + last.box.height - cell.box.y;
-      }
-      const authored = authoredById.get(cell.id);
-      let blocks = cell.blocks;
-      if (authored && authored.vAlign !== 'top' && blocks.length > 0) {
-        const insets = contentInsets(authored.margins, authored.borders);
-        // Content was placed relative to the first row; measure current content band.
-        let contentTop = Number.POSITIVE_INFINITY;
-        let contentBottom = Number.NEGATIVE_INFINITY;
-        for (const block of blocks) {
-          contentTop = Math.min(contentTop, block.box.y);
-          contentBottom = Math.max(contentBottom, block.box.y + block.box.height);
-        }
-        if (Number.isFinite(contentTop) && Number.isFinite(contentBottom)) {
-          const available = height - insets.top - insets.bottom - (contentBottom - contentTop);
-          // Reset any per-row shift by measuring from cell top + inset.
-          const desiredTop =
-            cell.box.y +
-            insets.top +
-            (available > 0 ? (authored.vAlign === 'center' ? available / 2 : available) : 0);
-          const dy = desiredTop - contentTop;
-          if (Math.abs(dy) > 0.001) {
-            blocks = shiftBlocks(blocks, dy);
-            for (const block of blocks) {
-              if (block.kind === 'paragraph') onAnchorShift?.(block.paragraphId, dy);
-            }
-          }
-        }
-      }
-      const finalizedCellBox = Object.freeze({
-        x: cell.box.x,
-        y: cell.box.y,
-        width: cell.box.width,
-        height,
-      });
-      if (
-        authored &&
-        anchorDeps &&
-        (span > 1 || (authored.vAlign !== 'top' && blocks.length > 0))
-      ) {
-        republishAnchoredParagraphsInBlocks(blocks, authored.blocks, finalizedCellBox, anchorDeps);
-      }
-      return {
-        ...cell,
-        ...(carried ? { vMergeContinue: false, paintInert: false } : {}),
-        rowSpan: span,
-        blocks,
-        box: { ...cell.box, height },
-      };
-    }),
-  }));
-
-  // Border grid from authored structure (same row/cell order as laid-out fragment rows).
-  // Header repeats use the same authored header row; match by cell id.
-  const gridRows: BorderGridCell[][] = expanded.map((row) =>
-    row.cells.map((cell) => {
-      const authored = authoredById.get(cell.id);
-      return {
-        gridColumn: cell.gridColumn,
-        gridSpan: cell.gridSpan,
-        vMergeContinue: cell.vMergeContinue,
-        borders: authored?.borders ?? {
-          top: { state: 'omitted' as const },
-          left: { state: 'omitted' as const },
-          bottom: { state: 'omitted' as const },
-          right: { state: 'omitted' as const },
-        },
-        mergeRowSpan: cell.rowSpan ?? 1,
-      };
-    })
-  );
-
-  const columnCount = structure.columnWidthsPt.length;
-  const tableBorders: TableBorderBox = structure.tableBorders;
-  const geometry: BorderGridGeometry = {
-    columnWidthsPt: structure.columnWidthsPt,
-    rowBands: expanded.map((row) => ({ y: row.box.y, height: row.box.height })),
-    cellBoxes: expanded.map((row) =>
-      row.cells.map((cell) => ({ width: cell.box.width, height: cell.box.height }))
-    ),
-  };
-  const resolved = resolveTableCellBorderGrid(
-    gridRows,
-    tableBorders,
-    columnCount,
-    geometry,
-    undefined,
-    ownershipBudget
-  );
-
-  return expanded.map((row, rowIndex) => ({
-    ...row,
-    cells: row.cells.map((cell, cellIndex) => {
-      const borders = resolved[rowIndex]![cellIndex]!;
-      if (cell.paintInert || cell.vMergeContinue) {
-        return { ...cell, borders: {}, paintInert: true };
-      }
-      return { ...cell, borders };
-    }),
-  }));
-}
 
 /**
  * A nested table inside a cell: laid out with its own geometry, no pagination, one
@@ -1787,16 +1569,15 @@ function emitNestedTable(
   const rawRows: TableRowFragmentRecord[] = [];
   let y = top;
   for (const [rowIndex, row] of structure.rows.entries()) {
-    const placed = layoutRowFragment(
+    const placed = layoutOnePassRow(
       row,
-      structure.columnWidthsPt,
+      structure,
       tableLeft,
       y,
-      false,
       depth,
       nestedFlowDeps,
-      structure.cellSpacingPt,
-      acceptVMergeSpansAt(vMergePlan, rowIndex)
+      false,
+      acceptVMergeSpansAt(vMergePlan, rowIndex, y)
     );
     rawRows.push(placed.record);
     y = placed.bottom;
@@ -1858,16 +1639,15 @@ export function layoutTableFragment(
   const rawRows: TableRowFragmentRecord[] = [];
   let y = top;
   for (const [rowIndex, row] of structure.rows.entries()) {
-    const placed = layoutRowFragment(
+    const placed = layoutOnePassRow(
       row,
-      structure.columnWidthsPt,
+      structure,
       left,
       y,
-      isHeaderRepeat(row),
       depth,
       deps,
-      structure.cellSpacingPt,
-      acceptVMergeSpansAt(vMergePlan, rowIndex)
+      isHeaderRepeat(row),
+      acceptVMergeSpansAt(vMergePlan, rowIndex, y)
     );
     rawRows.push(placed.record);
     y = placed.bottom;
